apps/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Python worker for Aqshara background jobs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.39.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.11.0",
  "pymupdf>=1.26.0",
  "python-docx>=1.2.0",
  "python-multipart>=0.0.20",
  "requests>=2.32.0",
  "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
  "mypy>=1.17.0",
  "pytest>=8.4.0",
  "pytest-cov>=6.2.0",
  "ruff>=0.12.0",
  "types-requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["worker"]
ignore_missing_imports = true
warn_return_any = false
disable_error_code = [
  "arg-type",
  "assignment",
  "attr-defined",
  "call-overload",
  "misc",
  "no-any-return",
  "return-value",
]
